Renaming folders in file tree for clarity.
