+++ src/api/admin/index.js
@@ -1,7 +1,7 @@
 const express = require('express')
 
-const adminGet = require('./get')
-const adminPost = require('./post')
+const adminGet = require('./adminGet')
+const adminPost = require('./adminPost')
 // const adminDelete = require('./delete')
 // const adminPut = require('./put')
 // const adminPatch = require('./patch')
